added CoolProp related solvent attributes
Inside the SolventData, two more attributes called 'inCoolProp' and
'NameinCoolProp' are added. 'inCoolProp' is a boolean, which is True if
the solvent's data are available in CoolProp and False if not.
'NameinCoolProp' gives the solvent's name that can be used in
CoolProp.'NameinCoolProp' is capital-sensitive. CoolProp can identify the
solvent only by its name. If the solvent is unavailable in CoolProp,
'NameinCoolProp' is None